Lab4/CloudAPI/apis/urls.py
# apis/urls.py
from django.urls import path
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from .views import CalculationView, home

urlpatterns = [
    path('calculator/', CalculationView.as_view(), name="calculator"),
    path('home/', home, name='home'),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
