Lab4/CloudAPI/requirements.txt
# Django core framework
Django

# Django REST framework
djangorestframework

# PostgreSQL database adapter for Python
psycopg2-binary

# Asynchronous server gateway interface references
asgiref

# Time zone and timezone-related utilities
pytz

# SQL parsing utilities
sqlparse

# Token based authentication
djangorestframework-simplejwt

# Middleware for handling Cross-Origin Resource Sharing (CORS) in Django applications`
django-cors-headers
